Simplify Dispatcher class hierarchy
After this patch there are two base classes:

lava.tool.dispatcher.Dispatcher, this is the old LavaNonLegacyDispatcher
lava_tool.dispatcher.Dispatcher, this is the old LavaDispatcher

BaseDispatcher got upgrade to have a run() method that dispatches the command
and exists. This is used to create lava.tool.main with new LavaDispatcher class
that imports commands from lava.commands namespace.